Change ts file to  CreateCategory.stories.tsx, change css
